+++ common/dct.h
@@ -31,7 +31,7 @@
               i==3 ? FIX8(0.9415) :\
               i==4 ? FIX8(1.2651) :\
               i==5 ? FIX8(1.1910) :0)
-static const int x264_dct8_weight_tab[64] = {
+static const uint16_t x264_dct8_weight_tab[64] = {
     W(0), W(3), W(4), W(3),  W(0), W(3), W(4), W(3),
     W(3), W(1), W(5), W(1),  W(3), W(1), W(5), W(1),
     W(4), W(5), W(2), W(5),  W(4), W(5), W(2), W(5),
@@ -48,7 +48,7 @@ static const int x264_dct8_weight_tab[64] = {
 #define W(i) (i==0 ? FIX8(3.125) :\
               i==1 ? FIX8(1.25) :\
               i==2 ? FIX8(0.5) :0)
-static const int x264_dct4_weight2_tab[16] = {
+static const uint16_t x264_dct4_weight2_tab[16] = {
     W(0), W(1), W(0), W(1),
     W(1), W(2), W(1), W(2),
     W(0), W(1), W(0), W(1),
@@ -62,7 +62,7 @@ static const int x264_dct4_weight2_tab[16] = {
               i==3 ? FIX8(0.88637) :\
               i==4 ? FIX8(1.60040) :\
               i==5 ? FIX8(1.41850) :0)
-static const int x264_dct8_weight2_tab[64] = {
+static const uint16_t x264_dct8_weight2_tab[64] = {
     W(0), W(3), W(4), W(3),  W(0), W(3), W(4), W(3),
     W(3), W(1), W(5), W(1),  W(3), W(1), W(5), W(1),
     W(4), W(5), W(2), W(5),  W(4), W(5), W(2), W(5),
